readme: increase size of example pvc
The previous size of 2Mi was too small for the provisioner I was testing
with (ceph rbd) and failed to format a file system leading to a difficult
to debug situation. Increase the example size to 1Gi as I don't know of
any back-end that would have a problem with that size, and it's still
very small.

Signed-off-by: John Mulligan <jmulligan@redhat.com>