api: extend smbservice to accept either pvc name or spec
Instead of having the user need to know two different CRDs for different
use cases, have one CR that can be configured in either way.
This change only adapts the existing logic for the new type. It does
not implement the logic needed to set up the PVC as part of the
SmbService.

Signed-off-by: John Mulligan <jmulligan@redhat.com>