makefile: add check rule & check-format & check-revive
The `check` rule is a top-level rule that calls other checks.
The `check-format` rule checks go code formatting.
The `check-revive` rule performs various code quality checks using the
revive tool configured with .revive.toml.

Using a similar pattern as kustomise and other codegen tools from the
sdk makefile, revive will be downloaded on demand if it is not already
on the $PATH.

Signed-off-by: John Mulligan <jmulligan@redhat.com>