workflows: fix podmanbuild workflow by setting a podman config
Previously, the podmanbuild target was failing due to running out of
space. This appears to be due to the "vfs" backend which is IIRC not the
recommended one. This sets "overlayfs" with the fuse overlay driver. In
addition, turn on host networking to see if it speeds things up any.

Signed-off-by: John Mulligan <jmulligan@redhat.com>