tests: add a new smbclient utils package
This is a test utility lib that helps run smbclient client commands
in a pod in the kubernetes cluster under test.
Currently, this simply shells out to kubectl. In the future we
may want to drop the kubectl dependency and use some apis, but this was
simpler to get started with.

Signed-off-by: John Mulligan <jmulligan@redhat.com>