integration tests: add a test case to verify the deployed image
This test case is intended to verify that the operator running in the
cluster is the expected one. The expected image can changed by setting
an environment variable. This test case can be skipped by setting
the environment variable to an empty string.
The intent is mainly for use in CI to ensure the operator running
in the cluster matches what was built for the test.

Signed-off-by: John Mulligan <jmulligan@redhat.com>