makefile: as part of the "build" compile the integration tests
Any errors in the test code can be found sooner by compiling them as
part of a build target in the makefile.
I left the resulting binary in the bin dir, along side the "actual"
operator binary. The other option is to simply send the result to
/dev/null as we only care that it compiles correctly. But may be the
binary will be more useful in the future?

Signed-off-by: John Mulligan <jmulligan@redhat.com>