workflows: set up a k3d registry and test locally built images
Previously, the integration tests run by the CI were always testing
the "latest" image on the quay io repo, not the changes that are part
of a PR, etc.
This change adds a k3d managed registry and builds the image as
part of the test-kubernetes job. The image is pushed into the k3d
registry and the tests are configured to consume and verify that
the image under test is the one we made as part of the PR/CI job.

Signed-off-by: John Mulligan <jmulligan@redhat.com>