integration tests: add test to verify resources are kept in one namespace
Previously, the operator created new resources in whatever namespace the
SmbShares were created in. This behavior has been changed so that
the operator keeps to one namespace, which is the one it is deployed in
by default. Add a test to verify this behavior.

Note that the secrets must be accessed by the pods and so secrets must
be created in the namespace where the pods will be. Generally, I expect
that in most use cases the SmbShares and SmbSecurityConfigs will also
be created in that namespace. If not, things could get confusing.
We'll see how this ends up working out over time.

Signed-off-by: John Mulligan <jmulligan@redhat.com>