kube utils: allow tests to access dynamic client
Add a public DynamicClientset function to the kube utils package so that
I can defer figuring out the best way to access SmbShare/etc resources
in the tests. There is probably a better way to do this, but it works
for now and the test utils aren't really meant to be PUBLIC public
anyway.

Signed-off-by: John Mulligan <jmulligan@redhat.com>