makefile: start treating vet as a check
This adds vet as a check in `make check`. Additionally, it moves
the run of vet in `make test` and `make manager` to after the build.
I wanted to remove it entirely from that line but decided not to make
such a big change. Vet needs to run after compile because it prints
a 'vet: ' prefix on compile errors [1], and this messes up tools -
specifically my vim quickfix. Doing the build first makes you fix
real compiler errors and then real vet errors found afterward get
emitted without the annoying prefix.

[1]: https://github.com/golang/go/issues/34142

Signed-off-by: John Mulligan <jmulligan@redhat.com>