smbclient utils: add a function to clear client caches
The CacheFlush is meant to ensure that no persistent caches remain
on the client. Currently the implementation is a hack using rm
because the client container lacks the `net` command. Once that
issue is resolved this could be changed to use something like
`net cache flush` or any other more integrated approach.

Signed-off-by: John Mulligan <jmulligan@redhat.com>