controllers: remove unused test file
Remove unused "unit" test stub file. We never added any actual tests to
it.
The team prefers a mix of more unit-test-level tests and fully fledged
integration tests running in a k8s cluster. Even if we needed "unit"
tests that ran the controller itself I'm not particularly interested in
the test framework that the sdk opted us into, so there's little reason
to keep this file around any more.

Signed-off-by: John Mulligan <jmulligan@redhat.com>