resources: fix failure to check returned errors from SetControllerReference
This also makes the error logging calls consistent between the calls to
SetControllerReference and the API calls to create/update/whatever the
resource. This eliminates the functions that existed to create a
resource type and then set the controller reference as it was easier
to inline these "two line" functions than plumb error handling through
everything.

Signed-off-by: John Mulligan <jmulligan@redhat.com>