revive: add new rules to revive's configuration
These rules are ones I find to be fairly obviously useful. Ones that
should be near to a no-brainer for this codebase.

I expect there could be some quibbling with the line length I've
selected here but as the most active dev on this project I figure that I
have accrued enough points to make an arbitrary choice here and there.

Signed-off-by: John Mulligan <jmulligan@redhat.com>