main: log the go version used along with code version
For debugging/provenance we added an early log line that includes the
version and commit id of the operator. To that, we add the version of go
used for the build. It's something I needed on a different project, its
fresh in my mind, and it's better to do this now when we don't need it
then scramble for it later.

Signed-off-by: John Mulligan <jmulligan@redhat.com>