workflows: require the use of go 1.16
Go 1.16 is the lowest currently supported version of go. However, the
github action 'actions/setup-go@v2' was defaulting to an older
unsupported Go version. Some of our newer dependencies require Go 1.16
or later so we now require it specifically in our yaml. Unfortunately,
this will mean we need to remember to update it ourselves when Go makes
a new release.

Signed-off-by: John Mulligan <jmulligan@redhat.com>