makefile: add a "DEVELOPER" var to isolate personal hackery
Now that the various makefile rules using kustomize are variable-ized,
we add a "shortcut" variable "DEVELOPER" which, when set to an non-empty
value sets customize to use "config/developer" for it's
kustomization.yaml file. Outside of the readme this dir is to be ignored
by git so that it can act as a private "playground" for developer
specific hacks and tomfoolery without concern of accidentally
influencing the state of the repo.

Signed-off-by: John Mulligan <jmulligan@redhat.com>