resources: prevent trivially changing backends
This uses the backend annotation to prevent simple changes to the
availability mode making a hash of the system. This is because
we don't have a mature system of reconciliation. Maybe some day.

Obviously, the annotation can be reset - but this is intentional. If a
user pays close enough attention to see this and they still break the
system they get to keep both parts. ;-)

Signed-off-by: John Mulligan <jmulligan@redhat.com>