integration tests: add a new suite of tests for reconciles
This new set of tests will be for mutating the state of CRs/etc
and verifying how the operator handles that.

The intial test verifies that changing the availability mode does NOT
change the state of the samba resources backing the share.
That will be the behavior until we can teach the operator to
gracefully handle that kind of change.

Signed-off-by: John Mulligan <jmulligan@redhat.com>