test utils: add a library for polling for conditions
The test code is going to need more than a few places were we poll or
wait for a condition to be true. This new small library adds a general
mechanism for doing that kind of thing.

The TryUntil function executes a condition function until it returns
true on an error. The interfaces let one build upon the condition
by adding optional control of the interval between probes and an
interface for controlling what is done before the poll function(s)
exits. The latter can be used to gather stats, update errors, etc.
Finally, there's a small helper type that adapts basic functions
into meeting the Poller interface.

Signed-off-by: John Mulligan <jmulligan@redhat.com>