integration tests: check for >=2 pods for ctdb+ad suite
Add support for the common smb share suites to ensure that more
than one pod exists and is ready. Use it for when we're testing
the full feature set of the ctdb clustering + AD membership with
DNS integration.

While were at it, split the check into a SetupTest function that
runs before every test case. This can help detect errors that occur
after/between test cases.

Signed-off-by: John Mulligan <jmulligan@redhat.com>