test utils: add dynamic resource renaming
It turns out that some of the "slowness" we seeing in the test suite is
due to reconcile taking extra long because of "old" resources from
previous test suites using the same names that the new suite uses and
the old resources are not cleaned up by k8s yet. This delays the
operator until the conflicting resource is gone.

This change allows us to add name suffixes to resources loaded from
files (the common case) as well as adding general rename support to
our interfaces. Later changes to the tests can use the suffix to
disambiguate resources across tests while continuing to let us reuse the
same files. Partly inspired by kustomize.

Signed-off-by: John Mulligan <jmulligan@redhat.com>