integration tests: add cleanup tests for clustered shares
In addition to adding the tests for clustered (ctdb) share instances,
the test case is modified in order to make it more reliable. The
previous version worked if you were lucky but most of the time the
pods were in Terminating state a long time - far longer than was
reasonable to have the test wait. It turns out that the pod will
be terminating a longer time than usual in two cases I saw:
the pod is deleted before it becomes ready, and/or when the
secrets and configmaps it mounts are deleted before the pod is
gone. So now we wait for all expected pods to be ready and then
only delete the SmbShare resource before waiting for pods to
go away, and only then do we delete the resources we created
from files.

Signed-off-by: John Mulligan <jmulligan@redhat.com>