docs: improve developer doc around setting up kustomization.yaml
The docs were not clear on the fact that the snippets of
kustomization.yaml were not the full file and that typically you'd
want to inherit from the standard resources. This would automatically
get created by the makefile if you invoked commands in the right order
but it certainly was not obvious.
Now, the makefile has an explicit target for initializing a typical
developer dir we can document it and the fact that our snippets
are meant to be appended.

Signed-off-by: John Mulligan <jmulligan@redhat.com>