resources: add a volKeeper type for wrangling volumes
The current implementation of pod construction relies on managing many
slices of volMount objects. Not only is this a bit verbose but it also
relies on a trick that the static checkers don't like. In order to
make things a bit less "loose" and more regular the very simple
volKeeper type is added as a wrapper around a slice of volMounts.
Future changes will use this to work with various groupings of
volume/mounts.

Signed-off-by: John Mulligan <jmulligan@redhat.com>