resources: dedicated module to handle OpenShift logic
When running over OpenShift cluster some extra resources need to be
configured once, prior to creation of SMB shares:
  - ServiceAccount
  - Role (SCC, Metrics)
  - RoleBinding (SCC, Metrics)

Those resources are not associated with a specific SMB share, and
therefore they are reconciled as part of the SmbCommonConfig live-cycle.
Thus, when running over OpenShift SmbCommonConfig becomes mandatory
settings (even if empty).

Signed-off-by: Shachar Sharon <ssharon@redhat.com>