config: default to latest samba operator image
Since using the Operator SDK to create the initial kustomization files
we've defaulted to 'controller:latest' in the base yamls. However, this
means one *must* use kustomize to set the image to something. Typically,
we've been doing this using the Makefile.
However, now that we've well established that the canonical image for
the operator is at 'quay.io/samba.org/samba-operator' there's really no
need to require the image to be overridden. Update the manager base
YAML to default to the canonical operator image repo - and update the
makefile to match.

Signed-off-by: John Mulligan <jmulligan@redhat.com>