integration tests: add new TestGroup type
The TestGroup will serve as a future replacement for our current
map-based approach to running multiple test suites. The plan is
to have something that is deterministic by default and somewhat
simple to use. It's clearly more code than the old function but
the idea is that the various files defining tests will only
need to interact with the obvious parts of this type.

Signed-off-by: John Mulligan <jmulligan@redhat.com>