integration tests: use TestGroups to organize integration tests
Now that we have a TestGroup type we use it to organize all our
suites. This should hopefully make adding new tests to existing
groups and even adding new test files simpler. Especially, as
it should now be OK to copy and paste and just change the names
rather than piggybacking on unrelated groups of tests.

Signed-off-by: John Mulligan <jmulligan@redhat.com>