integration tests: give unique names to smb share resources
In the smb share tests, it's better to have unique names for the
resources we create, in order to tie them back to the tests that
created them. This can also help avoid name conflicts between
test cases and across test runs.

Because we want to have unique names we also have to remove some of
the more "static" configuration input parameters. This is also a
good cleanup as we should not assume the operator will name the
server resource after the share. In the future, this should be
derived from the serverGroup status field. However, we don't want
to do too much in these patches.

Signed-off-by: John Mulligan <jmulligan@redhat.com>