integration tests: make the smb login test loop a function
Recent changes made the test to log in to a share more robust by
putting it in a retry loop. There are other cases were we want to
robustly try to log into a share, so first refactor the code into
a function that can be used outside of the basic test suite.

Signed-off-by: John Mulligan <jmulligan@redhat.com>