resources: avoid placing data vol in metrics container namespace
In order to limit the "attack surface" of the various pods in our
container we should try to limit the data volume(s) to only the
containers that need them. This change uses the new volKeeper filtering
mechanism to drop any data volumes from being available to the metrics
container. This avoids having to build a complex set of volumes
for the metrics container while avoiding exposing the data volume
directly to that container.

Signed-off-by: John Mulligan <jmulligan@redhat.com>