planner: update configuration for readonly and browsable params
When the SmbShare resource is updated the operator will consider
changes to the readonly and browsable options and translate
those updates into the configmap for the pod(s).

Note: only those params are updated. Everything else is ignored,
eventually we'll need an admission controller or something like
that so changes to the "read-only" fields are rejected.

Signed-off-by: John Mulligan <jmulligan@redhat.com>