api: add Group and GroupMode Scaling options
Allowing shares to be grouped together under one smbd "instance" can
increase the density and potentially lower the costs of having many
shares in one system. Currently, the CRD either disable grouping
completely or require explicit grouping, where the creator of the share
specifies what group name to use.
This is part of the design that has been long outstanding.

Signed-off-by: John Mulligan <jmulligan@redhat.com>