resources: use ownership metadata for finding sibling shares
When looking to group shares under one server instance we need to
compare the smbshares and the properties of the shares to see if the
shares are compatible. We can use the existing ownership metadata to
determine what shares already own the resources supporting the server in
order to determine what shares to compare to for compatibility.

Signed-off-by: John Mulligan <jmulligan@redhat.com>