resources: add Yield method to Result type
The Yield method will return true if the result's content indicates
that no further processing should be done.

The intent is to allow methods that return result types to stack. In
other words, a function can perform reconcile-level actions and return a
`Done` result in which case that "task" is complete or it can set an
error or `Requeue` cases that indicate the caller should "yield" its
control flow and return the result from the function it called.

Signed-off-by: John Mulligan <jmulligan@redhat.com>