resources: break up the Update reconcile function into parts
Using the Result.Yield method, we can more easily separate parts of
the rather large Update function into sub-routines each returning
`Done` if that sub-routine has nothing else to do or setting an error
or requesting a `Requeue`.

View with -w (ignore whitespace) as well as default to see that
most of this patch is moving chunks of Update into new functions.

Signed-off-by: John Mulligan <jmulligan@redhat.com>