tests/integration: Fix flaky TestScaleoutClusterSuite
TestScaleoutClusterSuite fails more frequently with the following error:

=== RUN   TestIntegration/reconciliation/scaleoutCluster/TestScaleoutClusterSuite
    reconcile_test.go:151:
        Error Trace:    reconcile_test.go:151
        Error:          Not equal:
                        expected: 3
                        actual  : 2
        Test:           TestIntegration/reconciliation/scaleoutCluster/TestScaleoutClusterSuite
        Messages:       Clustersize not as expected

Above check is to make sure that number of replicas within StatefulSet
reflects the updated SmbShare.Spec.Scaling.MinClusterSize. But an
immediate check on StatefulSet.Spec.Replicas might not always give us
the desired(updated) value.

Therefore we retry this check within a brief 3 seconds timeout on
account of any delay in field update. In addition, we at least wait
for the existence of extra pods corresponding to updated replica count.

Signed-off-by: Anoop C S <anoopcs@samba.org>