resources: add claimOwnership for adding add'l owners to resources
The new claimOwnership function is for extending the ownership of
resources supporting a server instances beyond the smb share
resources that triggered the initial resource creation. Once
created resources have multiple owners the won't be automatically
garbage collected in the case the first SmbShare is deleted.
Includes initial unit test.

Signed-off-by: John Mulligan <jmulligan@redhat.com>