resources: call claimOwnership when reconciling
When we reconcile changes to SmbShares we want to make sure that they
are additional owners of resources that support the server instance
the shares will be hosted in.
As we pass through the various stages of reconciliation call the
claimOwnership function to make sure the "child" resources are
owned by all resources that need them.

Signed-off-by: John Mulligan <jmulligan@redhat.com>