resources: add ownership.go for ownership utility funcs
Along with unit tests in ownership_test.go, the functions in
ownership.go will be used for working with k8s' owner metadata.
The owner metadata help k8s decide when object should be garbage
collected. An object may have >1 owner, but only one controller.

We plan on having multiple SmbShares own resources like configmaps,
deployments, or statefulsets. The first SmbShare to trigger the
resources creation will be the controller-owner but we'll want
to manage changes to the ownership & controllership as shares
are created and deleted.

Signed-off-by: John Mulligan <jmulligan@redhat.com>