resources: add transferOwnership for changing controller-owner
The k8s concept of resources ownership has a few additional flags that only
one owner can poses. In particular the "controller" is tagged such that
the operator's reconcile loop is triggered for that controller-owner.
For resources like the config map, the deployment (and its scale field),
and things of that nature we want at least one of our SmbShares to
be controller-owner. Therefore we add a new function to help transfer
controller-ownership from an smbshare that is getting deleted to
another share.
Includes some unit test coverage.

Signed-off-by: John Mulligan <jmulligan@redhat.com>