config: OpenShift's SCC for samba
When running over OpenShift cluster, samba pods requires a dedicated
SecurityContextConstraints in order to function properly. This SCC is
not as powerful as the privileged but it has stronger capabilities
then the default one. In particular, with respect to modification to the
underlying volume.

It is up to the user to deploy this SCC prior to operator's deployment.
Either via kustomization.yaml or via explicit command line:

    $ oc apply -f config/openshift/samba-scc.yaml

Signed-off-by: Shachar Sharon <ssharon@redhat.com>