resources: for share based resources add security & config labels
Add labels for "samba-operator.samba.org/common-config-from" and
"samba-operator.samba.org/security-config-from" when the server instance
is derived from a share using common config or security config CRs. The
values of the labels map back to the names of the resources.
This can be handy for quickly seeing what resources came from where and
when writing tests that need to associate a pod with a CR resource
beyond just the SmbShare name.

Signed-off-by: John Mulligan <jmulligan@redhat.com>