integration tests: add env var SMBOP_TEST_MIN_NODE_COUNT
Add yet another environment variable to inform the test suite about the
outer world. The variable SMBOP_TEST_MIN_NODE_COUNT should be set to the
number of nodes that typical pods may be scheduled on, the tests may
then use this value to determine if the cluster is in-spec or
out-of-spec for tests that require a certain number of nodes.

Signed-off-by: John Mulligan <jmulligan@redhat.com>