integration tests: check for node availability in scheduling tests
The tests for the node selector and affinity scheduling require at least
two nodes. Previously, the tests would simply fail if there weren't
enough nodes to run the test. This patch adds a check such that the test
is skipped if there aren't enough nodes - unless the
SMBOP_TEST_MIN_NODE_COUNT was specified. This variable acts as a double
check on the expected environment. If the number of available nodes is
less than that given by the env var the test is required to fail.

Signed-off-by: John Mulligan <jmulligan@redhat.com>