workflows: use symbolic name oldstable instead of a particular version
The actions/setup-go action has supported symbolic names of `stable` and
`oldstable` for a bit. Instead of having to constantly update the
workflow every time Go does a new release use the `oldstable` keyword in
our workflow so that we're always testing with the older stable release
of Go. We choose the older stable version because we do not want to
accidentally introduce something new that requires a feature of the
newer Go release.

Signed-off-by: John Mulligan <jmulligan@redhat.com>