makefile: add check-dockerfile-go-version
Add a new check-dockerfile-go-version makefile target and include it in
the check target. This runs the recently added script to validate the
version of golang used to build in container.

The downside of this approach is that PRs & other work unrelated to
Golang may periodically break when Go issues a new major release.
Let's try this for now and if it ends up getting really annoying we can
try to do some other form of automation to alert us to the need of
a version bump that doesn't interfere with unrelated work.

Signed-off-by: John Mulligan <jmulligan@redhat.com>