github: Update mergify config to replace deprecated attributes
!! Action Required !!

The configuration uses the deprecated merge_method attribute of the
queue action in one or more pull_request_rules. It must now be used
under the queue_rules configuration.
A brownout is planned on August 26th, 2024.
This option will be removed on September 23rd, 2024.
For more information: [1]

!! Action Required !!

The configuration uses the deprecated update_method attribute of the
queue action in one or more pull_request_rules. It must now be used
under the queue_rules configuration.
A brownout is planned on August 26th, 2024.
This option will be removed on September 23rd, 2024.
For more information: [1]

ref: https://docs.mergify.com/workflow/actions/queue/#parameters
[1] https://docs.mergify.com/configuration/file-format/#queue-rules

Signed-off-by: Anoop C S <anoopcs@samba.org>