go.mod: Bump Go version to 1.23.0
It is really unfortunate that golang has decided to come up with
an automation[1] to update `go` directive in go.mod from various
repositories including x/repos following the release cadence in
general. Another interesting and visible change is to explicitly
specify the minor version as a suffix to the major version. Thus
we are bound to update our `go` directive as we require many of
the golang x/repos.

[1] https://github.com/golang/go/issues/69095

Signed-off-by: Anoop C S <anoopcs@samba.org>