hack: Update controller-gen to v0.19.0
With go1.25, there's a breaking change[1] from golang.org/x/tools
pulled in by controller tools for which a fix is available in either
v0.24.1 (and above) or v0.25.1 (and above). This minimum version of
golang.org/x/tools is only included from controller tools v0.16.4.
We make use of this opportunity to fetch the latest version.

[1] https://github.com/golang/go/issues/74462

Signed-off-by: Anoop C S <anoopcs@samba.org>